RI-6371 move func to utils, made sort case insensitive
