+++ redisinsight/ui/src/utils/instance/index.ts
@@ -2,6 +2,7 @@ import getUrl from './getUrlInstance'
 
 export * from './instanceModules'
 export * from './instanceOptions'
+export * from './instanceNavigation'
 
 export {
   getUrl
